requirements.txt
fastapi
langchain-openai
python-dotenv
uvicorn
